+++ app/components/data/data.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from "@angular/core";
 
+import { FileModule } from "@batch-flask/ui";
 import { commonModules } from "app/common";
 import {
     DeleteContainerDialogComponent,
@@ -8,7 +9,6 @@ import {
 } from "app/components/data/action";
 import { DataSharedModule } from "app/components/data/shared";
 import { FileBrowseModule } from "app/components/file/browse";
-import { FileDetailsModule } from "app/components/file/details";
 import { DataContainerListComponent, FileGroupPreviewComponent } from "./browse";
 import {
     DataContainerConfigurationComponent,
@@ -33,7 +33,7 @@ const components = [
 @NgModule({
     declarations: components,
     exports: components,
-    imports: [...commonModules, DataSharedModule, FileBrowseModule, FileDetailsModule],
+    imports: [...commonModules, DataSharedModule, FileBrowseModule, FileModule],
     entryComponents: [
         DeleteContainerDialogComponent,
         FileGroupCreateFormComponent,
