+++ app/components/file/browse/file-browse.module.ts
@@ -2,7 +2,6 @@ import { NgModule } from "@angular/core";
 import { FileModule } from "@batch-flask/ui";
 import { commonModules } from "app/common";
 import { BlobFilesBrowserComponent } from "./blob-files-browser";
-import { FileExplorerModule } from "./file-explorer";
 import { NodeFileBrowseComponent } from "./node-file-browse.component";
 
 const components = [
@@ -10,7 +9,6 @@ const components = [
 ];
 
 const publicModules = [
-    FileExplorerModule,
 ];
 @NgModule({
     declarations: components,
