+++ src/@batch-flask/ui/file/file-explorer/file-explorer-tabs/file-explorer-tabs.component.ts
