+++ src/@batch-flask/ui/file/file-explorer/file-explorer.scss
