+++ src/@batch-flask/ui/file/file-explorer/file-table-view/file-table-view.scss
