+++ src/@batch-flask/ui/file/file-explorer/file-tree-view/file-tree-view-row/file-tree-view-row.component.ts
